OP-940 Fixed claim/GetClaims, updated IsoDateSerializer
