+++ OpenImis.ModulesV3/ClaimModule/Models/DsiInputModel.cs
@@ -1,12 +1,7 @@
 ﻿
 using Newtonsoft.Json;
-using OpenImis.ModulesV3.Helpers.Validators;
-using OpenImis.ModulesV3.Utils;
+using OpenImis.ModulesV3.Helpers;
 using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace OpenImis.ModulesV3.ClaimModule.Models
 {
