+++ OpenImis.ModulesV3/ClaimModule/Models/RegisterClaim/Claim.cs
@@ -1,5 +1,5 @@
 ﻿using Newtonsoft.Json;
-using OpenImis.ModulesV3.Utils;
+using OpenImis.ModulesV3.Helpers;
 using System;
 using System.Collections.Generic;
