+++ OpenImis.ModulesV3/Helpers/BusinessException.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net;
-using System.Threading.Tasks;
 
-namespace OpenImis.RestApi.Util.ErrorHandling
+namespace OpenImis.ModulesV3.Helpers
 {
     /// <summary>
     /// Exception for business related errors. It's error message will be included in the payload.
