+++ OpenImis.ModulesV3/Helpers/DateTimeFormats.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OpenImis.ModulesV3.Helpers
+{
+    public static class DateTimeFormats
+    {
+        public const string IsoDateFormat = "yyyy-MM-dd";
+        public const string IsoDateTimeFormat = "yyyy-MM-ddTHH:mm:ss";
+        public const string FileNameDateTimeFormat = "yyyy-MM-ddTHH-mm-ss";
+    }
+}
