+++ OpenImis.ModulesV3/PolicyModule/Models/PolicyRenewalModel.cs
@@ -1,8 +1,6 @@
 ﻿using Newtonsoft.Json;
-using OpenImis.ModulesV3.Utils;
+using OpenImis.ModulesV3.Helpers;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace OpenImis.ModulesV3.PolicyModule.Models
 {
@@ -14,14 +12,11 @@ namespace OpenImis.ModulesV3.PolicyModule.Models
         public string ReceiptNo { get; set; }
         public string ProductCode { get; set; }
         public float Amount { get; set; }
-
         [JsonConverter(typeof(IsoDateSerializer))]
         public DateTime Date { get; set; }
         public bool Discontinue { get; set; }
         public int PayerId { get; set; }
-
         public string ControlNumber { get; set; }
-
 
         public Policy GetPolicy()
         {
