+++ OpenImis.ModulesV3/Utils/LinqExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 
 namespace OpenImis.ModulesV3.Utils
 {
