+++ OpenImis.ModulesV3/Utils/PhotoUtils.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
 using Microsoft.Extensions.Logging;
 
 namespace OpenImis.ModulesV3.Utils
