+++ OpenImis.ModulesV3/Utils/TypeCast.cs
@@ -1,8 +1,5 @@
 ﻿using Newtonsoft.Json;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace OpenImis.ModulesV3.Utils
 {
