+++ OpenImis.RestApi/Controllers/V3/PolicyRenewalController.cs
@@ -7,9 +7,9 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using OpenImis.ModulesV3;
+using OpenImis.ModulesV3.Helpers;
 using OpenImis.ModulesV3.PolicyModule.Models;
 using OpenImis.ModulesV3.Utils;
-using OpenImis.RestApi.Util.ErrorHandling;
 using OpenImis.Security.Security;
 
 namespace OpenImis.RestApi.Controllers.V3
