+++ OpenImis.RestApi/Util/ErrorHandling/ErrorHandlerMiddleware.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
+using OpenImis.ModulesV3.Helpers;
 using System;
 using System.Net;
 using System.Threading.Tasks;
